+++ lib/commands.py
@@ -55,6 +55,7 @@ log = logging.getLogger(__name__)
 
 known_commands = {}
 ADDRESS_LENGTH = 25
+MAX_PAGE_SIZE = 500
 
 # Format output from lbrycrd to have consistently
 # named ditionary keys
@@ -555,9 +556,9 @@ class Commands:
     def sendclaimtoaddress(self, claim_id, destination, amount, tx_fee=None, change_addr=None, broadcast=True):
         claims = self.getnameclaims(raw=True, include_supports=False, claim_id=claim_id)
         if len(claims) > 1:
-            return {"success": False ,'reason':'more than one claim that matches'}
+            return {"success": False, 'reason': 'more than one claim that matches'}
         elif len(claims) == 0:
-            return {"success": False, 'reason':'claim not found'}
+            return {"success": False, 'reason': 'claim not found', 'claim_id': claim_id}
         else:
             claim = claims[0]
 
@@ -768,17 +769,14 @@ class Commands:
         self.network.send([('blockchain.address.subscribe', [address])], callback)
         return True
 
-    def validate_claim_signature_and_get_channel_name(self, claim, claim_address):
-        cert_id = claim.certificate_id
-        response = self.network.synchronous_get(('blockchain.claimtrie.getclaimbyid',
-                                                 [cert_id]))
-        if response:
-            certificate = smart_decode(response['value'])
-            if Commands._validate_signed_claim(claim, claim_address, certificate):
-                return True, response['name']
+    def validate_claim_signature_and_get_channel_name(self, claim, certificate_claim,
+                                                      claim_address):
+        certificate = smart_decode(certificate_claim['value'])
+        if Commands._validate_signed_claim(claim, claim_address, certificate):
+            return True, certificate_claim['name']
         return False, None
 
-    def parse_and_validate_claim_result(self, claim_result, raw=False):
+    def parse_and_validate_claim_result(self, claim_result, certificate=None, raw=False):
         if not claim_result or 'value' not in claim_result:
             return claim_result
 
@@ -797,10 +795,15 @@ class Commands:
         if decoded:
             claim_result['has_signature'] = False
             if decoded.has_signature:
+                if certificate is None:
+                    print_msg("fetching certificate to check claim signature")
+                    certificate = self.getclaimbyid(decoded.certificate_id)
+                    if not certificate:
+                        raise Exception('Certificate claim {} not found'.format(decoded.certificate_id))
                 claim_result['has_signature'] = True
                 claim_result['signature_is_valid'] = False
                 validated, channel_name = self.validate_claim_signature_and_get_channel_name(
-                    decoded, claim_result['address'])
+                    decoded, certificate, claim_result['address'])
                 claim_result['channel_name'] = channel_name
                 if validated:
                     claim_result['signature_is_valid'] = True
@@ -925,7 +928,7 @@ class Commands:
             depth = self.network.get_server_height() - height
         result = Commands._verify_proof(name, block_header['claim_trie_root'], response,
                                         height, depth)
-        return self.parse_and_validate_claim_result(result, raw)
+        return self.parse_and_validate_claim_result(result, raw=raw)
 
     @command('n')
     def getclaimbynameinchannel(self, uri, name, raw=False):
@@ -939,7 +942,7 @@ class Commands:
         for claim in channel_claims:
             if claim['name'] == name:
                 return claim
-        return {'error': 'claim not found'}
+        return {'success': False, 'error': 'claim not found', 'name': name}
 
     @command('wn')
     def getdefaultcertificate(self):
@@ -952,48 +955,229 @@ class Commands:
             return {'error': 'no default certificate configured'}
         return self.getclaimbyid(certificate_id)
 
+    @staticmethod
+    def prepare_claim_queries(start_position, query_size, channel_claim_infos):
+        queries = [tuple()]
+        names = {}
+        # a table of index counts for the sorted claim ids, including ignored claims
+        absolute_position_index = {}
+
+        block_sorted_infos = sorted(channel_claim_infos.iteritems(), key=lambda x: int(x[1][1]))
+        per_block_infos = {}
+        for claim_id, (name, height) in block_sorted_infos:
+            claims = per_block_infos.get(height, [])
+            claims.append((claim_id, name))
+            per_block_infos[height] = sorted(claims, key=lambda x: int(x[0], 16))
+
+        abs_position = 0
+
+        for height in sorted(per_block_infos.keys()):
+            for claim_id, name in per_block_infos[height]:
+                names[claim_id] = name
+                absolute_position_index[claim_id] = abs_position
+                if abs_position >= start_position:
+                    if len(queries[-1]) >= query_size:
+                        queries.append(tuple())
+                    queries[-1] += (claim_id,)
+                abs_position += 1
+        return queries, names, absolute_position_index
+
+    def iter_channel_claims_pages(self, queries, claim_positions, claim_names, certificate,
+                                  page_size=10):
+        # lbryum server returns a dict of {claim_id: (name, claim_height)}
+        # first, sort the claims by block height (and by claim id int value within a block).
+
+        # map the sorted claims into getclaimsbyids queries of query_size claim ids each
+
+        # send the batched queries to lbryum server and iteratively validate and parse
+        # the results, yield a page of results at a time.
+
+        # these results can include those where `signature_is_valid` is False. if they are skipped,
+        # page indexing becomes tricky, as the number of results isn't known until after having
+        # processed them.
+        # TODO: fix ^ in lbryschema
+
+        def iter_validate_channel_claims():
+            for claim_ids in queries:
+                batch_result = self.network.synchronous_get(("blockchain.claimtrie.getclaimsbyids", claim_ids))
+                for claim_id in claim_ids:
+                    claim = batch_result[claim_id]
+                    if claim['name'] == claim_names[claim_id]:
+                        formatted_claim = self.parse_and_validate_claim_result(claim, certificate)
+                        formatted_claim['absolute_channel_position'] = claim_positions[claim['claim_id']]
+                        yield format_amount_value(formatted_claim)
+                    else:
+                        print_msg("ignoring claim with name mismatch %s %s" % (claim['name'], claim['claim_id']))
+
+        yielded_page = False
+        results = []
+        for claim in iter_validate_channel_claims():
+            results.append(claim)
+
+            # if there is a full page of results, yield it
+            if len(results) and len(results) % page_size == 0:
+                yield results[-page_size:]
+                yielded_page = True
+
+        # if we didn't get a full page of results, yield what results we did get
+        if not yielded_page:
+            yield results
+
+    def get_channel_claims_page(self, channel_claim_infos, certificate, page, page_size=10):
+        page = page or 0
+        page_size = max(page_size, 1)
+        if page_size > MAX_PAGE_SIZE:
+            raise Exception("page size above maximum allowed")
+        start_position = (page - 1) * page_size
+        queries, names, claim_positions = self.prepare_claim_queries(start_position, page_size,
+                                                                     channel_claim_infos)
+        page_generator = self.iter_channel_claims_pages(queries, claim_positions, names,
+                                                        certificate, page_size=page_size)
+        upper_bound = len(claim_positions)
+        if not page:
+            return None, upper_bound
+        if start_position > upper_bound:
+            raise IndexError("claim %i greater than max %i" % (start_position, upper_bound))
+        return next(page_generator), upper_bound
+
+    def _handle_resolve_uri_response(self, parsed_uri, block_header, raw, resolution, page=0,
+                                     page_size=10):
+        result = {}
+        # parse an included certificate
+        if 'certificate' in resolution:
+            certificate_response = resolution['certificate']['result']
+            certificate_resolution_type = resolution['certificate']['resolution_type']
+            if certificate_resolution_type == "winning" and certificate_response:
+                if 'height' in certificate_response:
+                    height = certificate_response['height']
+                    depth = self.network.get_server_height() - height
+                    certificate_result = Commands._verify_proof(parsed_uri.name,
+                                                                block_header['claim_trie_root'],
+                                                                certificate_response,
+                                                                height, depth)
+                    result['certificate'] = self.parse_and_validate_claim_result(certificate_result,
+                                                                                 raw=raw)
+            elif certificate_resolution_type == "claim_id":
+                result['certificate'] = self.parse_and_validate_claim_result(certificate_response,
+                                                                             raw=raw)
+            elif certificate_resolution_type == "sequence":
+                result['certificate'] = self.parse_and_validate_claim_result(certificate_response,
+                                                                             raw=raw)
+            else:
+                print_stderr("unknown response type: %s" % certificate_resolution_type)
+
+            if 'certificate' in result:
+                certificate = result['certificate']
+                if 'unverified_claims_in_channel' in resolution:
+                    max_results = len(resolution['unverified_claims_in_channel'])
+                    max_pages = 1 + int((max_results - (max_results % page_size)) / page_size)
+                    result['claims_in_channel_pages'] = max_pages
+                else:
+                    result['claims_in_channel_pages'] = 0
+            else:
+                result['error'] = "claim not found"
+                result['success'] = False
+                result['uri'] = str(parsed_uri)
+
+        else:
+            certificate = None
+
+        # if this was a resolution for a name, parse the result
+        if 'claim' in resolution:
+            claim_response = resolution['claim']['result']
+            claim_resolution_type = resolution['claim']['resolution_type']
+            if claim_resolution_type == "winning" and claim_response:
+                if 'height' in claim_response:
+                    height = claim_response['height']
+                    depth = self.network.get_server_height() - height
+                    claim_result = Commands._verify_proof(parsed_uri.name,
+                                                                block_header['claim_trie_root'],
+                                                                claim_response,
+                                                                height, depth)
+                    result['claim'] = self.parse_and_validate_claim_result(claim_result,
+                                                                           certificate,
+                                                                           raw)
+            elif claim_resolution_type == "claim_id":
+                result['claim'] = self.parse_and_validate_claim_result(claim_response,
+                                                                       certificate,
+                                                                       raw)
+            elif claim_resolution_type == "sequence":
+                result['claim'] = self.parse_and_validate_claim_result(claim_response,
+                                                                       certificate,
+                                                                       raw)
+            else:
+                print_stderr("unknown response type: %s" % claim_resolution_type)
+
+        # if this was a resolution for a name in a channel make sure there is only one valid
+        # match
+        elif 'unverified_claims_for_name' in resolution and 'certificate' in result:
+            unverified_claims_for_name = resolution['unverified_claims_for_name']
+
+            channel_info = self.get_channel_claims_page(unverified_claims_for_name,
+                                                        result['certificate'], page=1)
+            claims_in_channel, upper_bound = channel_info
+
+            if len(claims_in_channel) > 1:
+                print_stderr("Multiple signed claims for the same name")
+            elif not claims_in_channel:
+                print_stderr("No valid claims for this name for this channel")
+            else:
+                result['claim'] = claims_in_channel[0]
+
+        # parse and validate claims in a channel iteratively into pages of results
+        elif 'unverified_claims_in_channel' in resolution and 'certificate' in result:
+            ids_to_check = resolution['unverified_claims_in_channel']
+            channel_info = self.get_channel_claims_page(ids_to_check, result['certificate'],
+                                                        page=page, page_size=page_size)
+            claims_in_channel, upper_bound = channel_info
+
+            if claims_in_channel:
+                result['claims_in_channel'] = claims_in_channel
+        elif 'error' not in result:
+            result['error'] = "claim not found"
+            result['success'] = False
+            result['uri'] = str(parsed_uri)
+
+        return result
+
     @command('n')
-    def getvalueforuri(self, uri, raw=False):
+    def getvalueforuri(self, uri, raw=False, page=0, page_size=10):
         """
         Resolve a LBRY URI
         """
 
-        try:
-            parsed_uri = parse_lbry_uri(uri)
-        except URIParseError as err:
-            return {'error': err.message}
-        result = {}
-        if parsed_uri.is_channel:
-            if parsed_uri.path:
-                result['claim'] = self.getclaimbynameinchannel(uri, parsed_uri.path, raw)
-            else:
-                result['claims_in_channel'] = self.getclaimsinchannel(uri, raw)
+        result = self.getvaluesforuris(raw, page, page_size, uri)
+        if uri in result:
+            return result[uri]
+        return result
 
-        if parsed_uri.claim_id and parsed_uri.name:
-            claim = self.getclaimbyid(parsed_uri.claim_id, raw)
-            if not parsed_uri.is_channel and claim['claim_id'] != parsed_uri.claim_id:
-                return {'error': 'certificate does not match uri'}
-            elif parsed_uri.is_channel and parsed_uri.path and claim['name'] != parsed_uri.name:
-                return {'error': 'channel claim does not match uri'}
-            if parsed_uri.is_channel:
-                result['certificate'] = claim
-            else:
-                result['claim'] = claim
-        elif parsed_uri.claim_sequence:
-            if parsed_uri.is_channel:
-                result['certificate'] = self.getnthclaimforname(parsed_uri.name, parsed_uri.claim_sequence, raw)
-            else:
-                result['claim'] = self.getnthclaimforname(parsed_uri.name, parsed_uri.claim_sequence, raw)
-        else:
-            if parsed_uri.is_channel:
-                result['certificate'] = self.getvalueforname(parsed_uri.name, raw)
-            else:
-                result['claim'] = self.getvalueforname(parsed_uri.name, raw)
-        for k in result:
-            if 'error' in result[k]:
-                return result[k]
-        if not result.get('claim', False) and not result.get('certificate', False) and not result.get('claims_in_channel', False) and not result.get('error', False):
-            return {'error': 'claim not found'}
+    @command('n')
+    def getvaluesforuris(self, raw=False, page=0, page_size=10, *uris):
+        """
+        Resolve a LBRY URI
+        """
+
+        page = int(page)
+        page_size = int(page_size)
+        uris_to_send = ()
+
+        for uri in uris:
+            try:
+                parse_lbry_uri(uri)
+                uris_to_send += (str(uri), )
+            except URIParseError as err:
+                return {'error': err.message}
+
+        height = self.network.get_local_height() - RECOMMENDED_CLAIMTRIE_HASH_CONFIRMS + 1
+        block_header = self.network.blockchain.read_header(height)
+        block_hash = self.network.blockchain.hash_header(block_header)
+        response = self.network.synchronous_get(('blockchain.claimtrie.getvaluesforuris',
+                                                 (block_hash, ) + uris_to_send))
+        result = {}
+        for uri, resolution in response.iteritems():
+            result[uri] = self._handle_resolve_uri_response(parse_lbry_uri(str(uri)), block_header,
+                                                            raw, resolution, page=page,
+                                                            page_size=page_size)
         return result
 
     @command('n')
@@ -1017,7 +1201,7 @@ class Commands:
                 response['signature'] = decoded_claim.signature
                 response['certificate'] = self.getclaimbyid(decoded_claim.certificate_id)
                 validated, channel_name = self.validate_claim_signature_and_get_channel_name(
-                    decoded_claim, raw_claim['address'])
+                    decoded_claim, response['certificate'], raw_claim['address'])
                 response['channel_name'] = channel_name
                 if validated:
                     response['signature_is_valid'] = True
@@ -1037,7 +1221,7 @@ class Commands:
 
         out = self.network.synchronous_get(('blockchain.claimtrie.getclaimsintx', [txid]))
         result = format_amount_value(format_lbrycrd_keys(out, raw))
-        return self.parse_and_validate_claim_result(result, raw)
+        return self.parse_and_validate_claim_result(result, raw=raw)
 
     @command('n')
     def getclaimbyoutpoint(self, txid, nout, raw=False):
@@ -1049,7 +1233,7 @@ class Commands:
         claims = format_amount_value(format_lbrycrd_keys(out, raw))
         for claim in claims:
             if claim['nout'] == nout:
-                return self.parse_and_validate_claim_result(claim, raw)
+                return self.parse_and_validate_claim_result(claim, raw=raw)
 
     @command('n')
     def getclaimsforname(self, name, raw=False):
@@ -1062,7 +1246,7 @@ class Commands:
         claims = format_amount_value(result['claims'])
         claims_for_return = []
         for claim in claims:
-            claims_for_return.append(self.parse_and_validate_claim_result(claim, raw))
+            claims_for_return.append(self.parse_and_validate_claim_result(claim, raw=raw))
         result['claims'] = claims_for_return
         return result
 
@@ -1075,7 +1259,7 @@ class Commands:
         result = self.network.synchronous_get(('blockchain.claimtrie.getclaimssignedbyid',
                                                [claim_id]))
         claims = format_amount_value(result)
-        return [self.parse_and_validate_claim_result(claim, raw) for claim in claims]
+        return [self.parse_and_validate_claim_result(claim, raw=raw) for claim in claims]
 
     @command('n')
     def getclaimsinchannel(self, uri, raw=False):
@@ -1097,7 +1281,7 @@ class Commands:
                                                  [parsed.name]))
         if claims:
             result = format_amount_value(claims)
-            return [self.parse_and_validate_claim_result(claim, raw) for claim in result]
+            return [self.parse_and_validate_claim_result(claim, raw=raw) for claim in result]
         return []
 
     @command('n')
@@ -1151,7 +1335,7 @@ class Commands:
 
         result = self.network.synchronous_get(('blockchain.claimtrie.getclaimbyid', [claim_id]))
         claims = format_amount_value(result)
-        return self.parse_and_validate_claim_result(claims, raw)
+        return self.parse_and_validate_claim_result(claims, raw=raw)
 
     @command('n')
     def getnthclaimforname(self, name, n, raw=False):
@@ -1162,7 +1346,7 @@ class Commands:
         result = self.network.synchronous_get(('blockchain.claimtrie.getnthclaimforname',
                                                [name, n]))
         claims = format_amount_value(result)
-        return self.parse_and_validate_claim_result(claims, raw)
+        return self.parse_and_validate_claim_result(claims, raw=raw)
 
     @command('w')
     def getnameclaims(self, raw=False, include_abandoned=False, include_supports=True,
@@ -1176,7 +1360,7 @@ class Commands:
         claims = format_amount_value(result)
         name_claims = []
         for claim in claims:
-            parsed = self.parse_and_validate_claim_result(claim, raw)
+            parsed = self.parse_and_validate_claim_result(claim, raw=raw)
             if claim_id is not None and parsed['claim_id'] != claim_id:
                 continue
             if txid is not None and nout is not None:
@@ -1200,7 +1384,7 @@ class Commands:
             try:
                 decoded = smart_decode(claim['value'])
                 if decoded.is_certificate:
-                    cert_result = self.parse_and_validate_claim_result(claim, raw)
+                    cert_result = self.parse_and_validate_claim_result(claim, raw=raw)
                     if self.cansignwithcertificate(cert_result['claim_id']):
                         cert_result['can_sign'] = True
                     else:
@@ -1365,7 +1549,7 @@ class Commands:
     @command('wpn')
     def updateclaimsignature(self, name, amount=None, claim_id=None, certificate_id=None):
         """
-        Update name claim signature
+        Update an unsigned claim with a signature
         """
 
         claim_value = None
@@ -1384,13 +1568,16 @@ class Commands:
         claim = smart_decode(claim_value)
         if certificate_id is None:
             certificate_id = claim.certificate_id
+            certificate = self.getclaimbyid(certificate_id)
+            if not certificate:
+                raise Exception('Certificate claim {} not found'.format(decoded.certificate_id))
         if not self.cansignwithcertificate(certificate_id):
             return {
                 'error': ('can update claim for lbry://{}#{}, but the signing key is '
                           'missing for certificate {}').format(name, claim_id, certificate_id)
             }
         validated, channel_name = self.validate_claim_signature_and_get_channel_name(
-            claim, claim_address)
+            claim, certificate, claim_address)
         if validated:
             return {
                 'error': 'lbry://{}#{} has a valid signature already'.format(name, claim_id)
@@ -1751,9 +1938,9 @@ class Commands:
         claims = self.getnameclaims(raw=True, include_abandoned=False, include_supports=True,
                                     claim_id=claim_id, txid=txid, nout=nout)
         if len(claims) > 1:
-            return {"success": False ,'reason':'more than one claim that matches'}
+            return {"success": False, 'reason': 'more than one claim that matches'}
         elif len(claims) == 0:
-            return {"success": False, 'reason':'claim not found'}
+            return {"success": False, 'reason': 'claim not found', 'claim_id': claim_id}
         else:
             claim = claims[0]
 
@@ -1840,6 +2027,8 @@ command_options = {
     'claim_addr':  (None, "--claim_addr",  "Address where claims are sent."),
     'broadcast':   (None, "--broadcast",   "if True, broadcast the transaction"),
     'raw':      ("-r", "--raw", "if True, don't decode claim values"),
+    'page':     ("-p", "--page", "page number"),
+    'page_size': ("-s", "--page_size", "page size"),
     'claim_id': (None, "--claim_id", "claim id"),
     'txid': ("-t", "--txid", "txid"),
     'nout': ("-n", "--nout", "nout"),