+++ CHANGELOG.md
@@ -13,7 +13,8 @@ labeled as 2.7.1. Subsequent releases will follow
       
 ### Changed
   * Change 'nothing to resolve' error to 'claim not found' used in other places
-  *
+  * Move uri resolution logic to lbryum-server, validate response
+  * Support batched uri resolution
   
 ### Fixed
   * Fixed abandon command